+++ controllers/post-controller.js
@@ -6,17 +6,17 @@ async function post(req,res,next){
     
     const {title,description,blogData}  =req.body
     if(!title){
-        return res.send("title is required")
+        return res.status(400).send("title is required")
     }
     if(!description){
-        return res.send("description is required")
+        return res.status(400).send("description is required")
     }
     if(!blogData){
-        return res.send("post information is required")
+        return res.status(400).send("post information is required")
     }
 
     if(title.length > 200 || description.length > 300 || blogData.length > 2000){
-        return res.send("make sure title - 200 words, description - 300 words ,post - 2000 words")
+        return res.status(400).send("make sure title - 200 words, description - 300 words ,post - 2000 words")
     }
 
     const data = {
@@ -30,7 +30,7 @@ async function post(req,res,next){
     await backend.save()
     
     console.log("this is the middle ware for the blog post ")
-    res.send("new post has been created")
+    res.status(201).send("new post has been created")
     next()
 } 
 
@@ -40,26 +40,26 @@ async function post(req,res,next){
 async function postDelete(req,res,next){
    console.log("lets see if this is working")
     try{
-
-        if(req.body.blogId.length !=24){
-            return res.send("post not found")
+        
+        if(!req.body.blogId.length || req.body.blogId.length !=24){
+            return res.status(400).send("request not valid")
         }
         const postCreator = await blog.findOne({_id:req.body.blogId})
         if(postCreator == null){
-            return res.send("post not found")
+            return res.status(400).send("post not found")
         }
         console.log(postCreator.createdBy)
         console.log(req.body.user.name)
         if(postCreator.createdBy != req.body.user.name)
         {
-            return res.send("not authorized")
+            return res.status(403).send("not authorized")
 
 
         }
         
         
         await blog.deleteOne({_id:req.body.blogId})
-        return res.send("succefully deleted the post")
+        return res.status(200).send("succefully deleted the post")
 
     }
     catch(err){
@@ -74,16 +74,16 @@ async function postDelete(req,res,next){
 async function postUpdate(req,res,next){
     console.log("let's see if this is working postUpdate")
     try{
-        if(req.body.postId.length != 24){
-            return res.send("not found")
+        if(!req.body.postId || req.body.postId.length != 24){
+            return res.status(400).send("reqeuest body not valid")
         }
         const post = await blog.findOne({_id:req.body.postId})
         
         if(post == null){
-            return res.send("post not found")
+            return res.status(404).send("post not found")
         }
         if(post.createdBy != req.body.user.name){
-            return res.send("not authorized")
+            return res.status(403).send("not authorized")
         }
         
         const {title,blogData,description} = req.body
@@ -104,7 +104,7 @@ async function postUpdate(req,res,next){
 
         post.save()
         console.log(post)
-        return res.send("post updated successfully")
+        return res.status(200).send("post updated successfully")
     }catch(err){
         console.log(err)
     }
@@ -123,7 +123,7 @@ async function getSinglePost(req,res,next){
     const post = await blog.findById({_id:req.params.id})
     const something = "value"
     
-    res.send(something)
+    res.status(200).send(something)
     
     next()
 }
